refactor(#800): split cloud_sync.rs by moving tests to sibling
Move 2,586-line cloud_sync.rs into cloud_sync/ directory:

- cloud_sync/mod.rs:   1,281 LOC (production code)
- cloud_sync/tests.rs: 1,301 LOC (unit tests)

Both well below the 2,500 LOC limit. No behavior changes.

`cargo test -p budi-core --lib cloud_sync` green (43 passed).
